requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
mediapipe>=0.10.8
scikit-learn>=1.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
